Crate amimono

Amimono is a library-level modular monolith framework for building cloud-native applications with Rust that is lightweight and flexible.

With Amimono your application is broken up into components, which are then collected into jobs that can be run as independent workloads. The core Amimono runtime handles service discovery and component dispatch, and optional functionality such as the RPC subsystem makes it easy to define new components that can be used throughout the application.

Modules

config

Types and other definitions for defining Amimono applications.

rpc

Subsystem for building components with an RPC interface.

runtime

The entry point to the Amimono runtime.

Macros

rpc_ops

A helper macro for defining RPC components.

Functions

entry

The main Amimono entry point.