amimono

Module config

Types and other definitions for defining Amimono applications.

It’s important that AppConfigs be constructed deterministically so that all workloads can use the same value. In the future this may be encouraged by allowing configuration code to be written with const fns, but this is not currently implemented.

Example

Most config values are constructed with builders such as AppBuilder and JobBuilder, however ComponentConfig is meant to be constructed directly. The suggested way to organize your configuration code is to have one function per component that returns a ComponentConfig, and then a single function close to your main() function that assembles these into an AppConfig. This might look as follows:

use amimono::config::{AppBuilder, AppConfig, JobBuilder};

use crate::backend;
use crate::frontend;

pub fn configure() -> AppConfig {
    AppBuilder::new()
        .add_job(
            JobBuilder::new()
                .with_label("backend")
                .add_component(backend::emailer::component())
                .add_component(backend::accounts::component())
                .add_component(backend::images::component())
                .add_component(backend::orders::component())
                .build()
        )
        .add_job(
            JobBuilder::new()
                .with_label("cache")
                .add_component(backend::cache::component())
                .build()
        )
        .add_job(
            JobBuilder::new()
                .with_label("frontend")
                .add_component(frontend::component())
                .build()
        )
        .build()
}

However you are free to organize things in whatever way you prefer.

Structs

AppBuilder

A helper for constructing an AppConfig.

AppConfig

A fully configured application.

ComponentConfig

The configuration for a single component.

JobBuilder

A helper for constructing a JobConfig.

JobConfig

A fully configured job.

Enums

Binding

An allocated binding.

BindingType

A request for a type of binding.